+++ yasli/BinaryIArchive.h
@@ -24,7 +24,7 @@ public:
   bool operator()(unsigned short& value, const char* name = "", const char* label = 0);
   bool operator()(int& value, const char* name = "", const char* label = 0);
   bool operator()(unsigned int& value, const char* name = "", const char* label = 0);
-  bool operator()(__int64& value, const char* name = "", const char* label = 0);
+  bool operator()(long long& value, const char* name = "", const char* label = 0);
 
   bool operator()(signed char& value, const char* name = "", const char* label = 0);
   bool operator()(unsigned char& value, const char* name = "", const char* label = 0);
